Refactor route visitor
