+++ projects/gateway/pkg/translator/http.go
@@ -3,30 +3,17 @@ package translator
 import (
 	"context"
 	"fmt"
-	"strings"
-
-	"github.com/solo-io/gloo/projects/gateway/pkg/defaults"
-	matchersv1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1/core/matchers"
 	"github.com/solo-io/go-utils/hashutils"
 
 	"github.com/solo-io/go-utils/errors"
 
-	"github.com/solo-io/solo-kit/pkg/api/v1/resources"
 	"k8s.io/apimachinery/pkg/labels"
-
-	"github.com/gogo/protobuf/proto"
 
 	v1 "github.com/solo-io/gloo/projects/gateway/pkg/api/v1"
 	gloov1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
-	glooutils "github.com/solo-io/gloo/projects/gloo/pkg/utils"
 	"github.com/solo-io/go-utils/contextutils"
-	"github.com/solo-io/solo-kit/pkg/api/v1/resources/core"
 	"github.com/solo-io/solo-kit/pkg/api/v2/reporter"
 )
-
-// Reserved value for route table namespace selection.
-// If a selector contains this value in its 'namespace' field, we match route tables from any namespace
-const allNamespaceRouteTableSelector = "*"
 
 var (
 	NoVirtualHostErr = func(vs *v1.VirtualService) error {
@@ -75,7 +62,6 @@ func (t *HttpTranslator) GenerateListeners(ctx context.Context, snap *v1.ApiSnap
 	return result
 }
 
-// TODO(marco): validate SNI domains?
 // Errors will be added to the report object.
 func validateVirtualServiceDomains(gateway *v1.Gateway, virtualServices v1.VirtualServiceList, reports reporter.ResourceReports) {
 
@@ -264,8 +250,8 @@ func VirtualHostName(vs *v1.VirtualService) string {
 func convertRoutes(vs *v1.VirtualService, tables v1.RouteTableList, reports reporter.ResourceReports) ([]*gloov1.Route, error) {
 	var routes []*gloov1.Route
 	for _, r := range vs.GetVirtualHost().GetRoutes() {
-		rv := &routeVisitor{tables: tables}
-		mergedRoutes, err := rv.convertRoute(vs, r, reports)
+
+		mergedRoutes, err := NewRouteVisitor(vs, tables, reports).ConvertRoute(r)
 		if err != nil {
 			return nil, err
 		}
@@ -278,279 +264,4 @@ func convertRoutes(vs *v1.VirtualService, tables v1.RouteTableList, reports repo
 		routes = append(routes, mergedRoutes...)
 	}
 	return routes, nil
-}
-
-// converts a tree of gateway Routes into a list of Gloo routes
-type routeVisitor struct {
-	ctx     context.Context
-	tables  v1.RouteTableList
-	visited v1.RouteTableList
-}
-
-func (rv *routeVisitor) convertRoute(ownerResource resources.InputResource, ours *v1.Route, reports reporter.ResourceReports) ([]*gloov1.Route, error) {
-	matchers := []*matchersv1.Matcher{defaults.DefaultMatcher()}
-	if len(ours.Matchers) > 0 {
-		matchers = ours.Matchers
-	}
-
-	route := &gloov1.Route{
-		Matchers: matchers,
-		Options:  ours.Options,
-	}
-	switch action := ours.Action.(type) {
-	case *v1.Route_RedirectAction:
-		route.Action = &gloov1.Route_RedirectAction{
-			RedirectAction: action.RedirectAction,
-		}
-	case *v1.Route_DirectResponseAction:
-		route.Action = &gloov1.Route_DirectResponseAction{
-			DirectResponseAction: action.DirectResponseAction,
-		}
-	case *v1.Route_RouteAction:
-		route.Action = &gloov1.Route_RouteAction{
-			RouteAction: action.RouteAction,
-		}
-	case *v1.Route_DelegateAction:
-		return rv.convertDelegateAction(ownerResource, ours, reports)
-	}
-	return []*gloov1.Route{route}, nil
-}
-
-var (
-	matcherCountErr     = errors.New("invalid route: routes with delegate actions must omit or specify a single matcher")
-	missingPrefixErr    = errors.New("invalid route: routes with delegate actions must use a prefix matcher")
-	invalidPrefixErr    = errors.New("invalid route: route table matchers must begin with the prefix of their parent route's matcher")
-	hasHeaderMatcherErr = errors.New("invalid route: routes with delegate actions cannot use header matchers")
-	hasMethodMatcherErr = errors.New("invalid route: routes with delegate actions cannot use method matchers")
-	hasQueryMatcherErr  = errors.New("invalid route: routes with delegate actions cannot use query matchers")
-	delegationCycleErr  = func(cycleInfo string) error {
-		return errors.Errorf("invalid route: delegation cycle detected: %s", cycleInfo)
-	}
-
-	noDelegateActionErr = errors.New("internal error: convertDelegateAction() called on route without delegate action")
-
-	routeTableMissingWarning = func(ref core.ResourceRef) string {
-		return fmt.Sprintf("route table %v.%v missing", ref.Namespace, ref.Name)
-	}
-	noMatchingRouteTablesWarning    = "no route table matches the given selector"
-	invalidRouteTableForDelegateErr = func(delegatePrefix, pathString string) error {
-		return errors.Wrapf(invalidPrefixErr, "required prefix: %v, path: %v", delegatePrefix, pathString)
-	}
-	missingRefAndSelectorWarning = func(res resources.InputResource) string {
-		ref := res.GetMetadata().Ref()
-		return fmt.Sprintf("cannot determine delegation target for %T %s.%s: you must specify a route table "+
-			"either via a resource reference or a selector", res, ref.Namespace, ref.Name)
-	}
-)
-
-func (rv *routeVisitor) convertDelegateAction(routingResource resources.InputResource, route *v1.Route, reports reporter.ResourceReports) ([]*gloov1.Route, error) {
-	delegate := route.GetDelegateAction()
-	if delegate == nil {
-		return nil, noDelegateActionErr
-	}
-
-	delegatePrefix, err := getDelegateRoutePrefix(route)
-	if err != nil {
-		return nil, err
-	}
-
-	// Determine the route tables to delegate to
-	var routeTables v1.RouteTableList
-	if routeTableRef := getRouteTableRef(delegate); routeTableRef != nil {
-		// missing refs should only result in a warning
-		// this allows resources to be applied asynchronously
-		routeTable, err := rv.tables.Find((*routeTableRef).Strings())
-		if err != nil {
-			reports.AddWarning(routingResource, routeTableMissingWarning(*routeTableRef))
-			return nil, nil
-		}
-		routeTables = v1.RouteTableList{routeTable}
-	} else if rtSelector := delegate.GetSelector(); rtSelector != nil {
-		routeTables = routeTablesForSelector(rv.tables, rtSelector, routingResource.GetMetadata().Namespace)
-
-		if len(routeTables) == 0 {
-			reports.AddWarning(routingResource, noMatchingRouteTablesWarning)
-			return nil, nil
-		}
-	} else {
-		reports.AddWarning(routingResource, missingRefAndSelectorWarning(routingResource))
-		return nil, nil
-	}
-
-	// If any of the matching route tables has already been visited, that means we have a delegation cycle.
-	for _, visited := range rv.visited {
-		for _, toVisit := range routeTables {
-			if toVisit == visited {
-				return nil, delegationCycleErr(
-					buildCycleInfoString(append(append(v1.RouteTableList{}, rv.visited...), toVisit)),
-				)
-			}
-		}
-	}
-
-	var delegatedRoutes []*gloov1.Route
-	for _, routeTable := range routeTables {
-
-		// Create a new visitor to visit the current route table
-		// Copy over all route tables that have already been visited by the parent visitor and add the current one,
-		// so we can detect delegation cycles.
-		subRv := &routeVisitor{tables: rv.tables}
-		for _, vis := range rv.visited {
-			subRv.visited = append(subRv.visited, vis)
-		}
-		subRv.visited = append(subRv.visited, routeTable)
-
-		for _, routeTableRoute := range routeTable.Routes {
-
-			// clone route since we mutate
-			routeTableRoute := proto.Clone(routeTableRoute).(*v1.Route)
-
-			merged, err := mergeRoutePlugins(routeTableRoute.GetOptions(), route.GetOptions())
-			if err != nil {
-				// should never happen
-				return nil, errors.Wrapf(err, "internal error: merging route plugins from parent to delegated route")
-			}
-			routeTableRoute.Options = merged
-
-			if err := isRouteTableValidForDelegatePrefix(delegatePrefix, routeTableRoute); err != nil {
-				reports.AddError(routingResource, err)
-				continue
-			}
-
-			subRoutes, err := subRv.convertRoute(routeTable, routeTableRoute, reports)
-			if err != nil {
-				return nil, err
-			}
-			for _, sub := range subRoutes {
-				if err := appendSource(sub, routeTable); err != nil {
-					// should never happen
-					return nil, err
-				}
-				delegatedRoutes = append(delegatedRoutes, sub)
-			}
-		}
-	}
-
-	glooutils.SortRoutesByPath(delegatedRoutes)
-
-	return delegatedRoutes, nil
-}
-
-func getDelegateRoutePrefix(route *v1.Route) (string, error) {
-	switch len(route.GetMatchers()) {
-	case 0:
-		return defaults.DefaultMatcher().GetPrefix(), nil
-	case 1:
-		matcher := route.GetMatchers()[0]
-		var prefix string
-		if len(matcher.GetHeaders()) > 0 {
-			return prefix, hasHeaderMatcherErr
-		}
-		if len(matcher.GetMethods()) > 0 {
-			return prefix, hasMethodMatcherErr
-		}
-		if len(matcher.GetQueryParameters()) > 0 {
-			return prefix, hasQueryMatcherErr
-		}
-		if matcher.GetPathSpecifier() == nil {
-			return defaults.DefaultMatcher().GetPrefix(), nil // no path specifier provided, default to '/' prefix matcher
-		}
-		prefix = matcher.GetPrefix()
-		if prefix == "" {
-			return prefix, missingPrefixErr
-		}
-		return prefix, nil
-	default:
-		return "", matcherCountErr
-	}
-}
-
-func isRouteTableValidForDelegatePrefix(delegatePrefix string, routeTable *v1.Route) error {
-	for _, match := range routeTable.Matchers {
-		// ensure all subroutes in the delegated route table match the parent prefix
-		if pathString := glooutils.PathAsString(match); !strings.HasPrefix(pathString, delegatePrefix) {
-			return invalidRouteTableForDelegateErr(delegatePrefix, pathString)
-		}
-	}
-	return nil
-}
-
-// Handles new and deprecated format for referencing a route table
-// TODO: remove this function when we remove the deprecated fields from the API
-func getRouteTableRef(delegate *v1.DelegateAction) *core.ResourceRef {
-	if delegate.Namespace != "" || delegate.Name != "" {
-		return &core.ResourceRef{
-			Namespace: delegate.Namespace,
-			Name:      delegate.Name,
-		}
-	}
-	return delegate.GetRef()
-}
-
-func routeTablesForSelector(routeTables v1.RouteTableList, selector *v1.RouteTableSelector, ownerNamespace string) v1.RouteTableList {
-	type nsSelectorType int
-	const (
-		// Match route tables in the owner namespace
-		owner nsSelectorType = iota
-		// Match route tables in all namespaces watched by Gloo
-		all
-		// Match route tables in the specified namespaces
-		list
-	)
-
-	nsSelector := owner
-	if len(selector.Namespaces) > 0 {
-		nsSelector = list
-	}
-	for _, ns := range selector.Namespaces {
-		if ns == allNamespaceRouteTableSelector {
-			nsSelector = all
-		}
-	}
-
-	var labelSelector labels.Selector
-	if len(selector.Labels) > 0 {
-		labelSelector = labels.SelectorFromSet(selector.Labels)
-	}
-
-	var matchingRouteTables v1.RouteTableList
-	for _, candidate := range routeTables {
-
-		// Check whether labels match
-		if labelSelector != nil {
-			rtLabels := labels.Set(candidate.Metadata.Labels)
-			if !labelSelector.Matches(rtLabels) {
-				continue
-			}
-		}
-
-		// Check whether namespace matches
-		nsMatches := false
-		switch nsSelector {
-		case all:
-			nsMatches = true
-		case owner:
-			nsMatches = candidate.Metadata.Namespace == ownerNamespace
-		case list:
-			for _, ns := range selector.Namespaces {
-				if ns == candidate.Metadata.Namespace {
-					nsMatches = true
-				}
-			}
-		}
-
-		if nsMatches {
-			matchingRouteTables = append(matchingRouteTables, candidate)
-		}
-	}
-
-	return matchingRouteTables
-}
-
-func buildCycleInfoString(routeTables v1.RouteTableList) string {
-	var visitedTables []string
-	for _, rt := range routeTables {
-		visitedTables = append(visitedTables, fmt.Sprintf("[%s]", rt.Metadata.Ref().Key()))
-	}
-	return strings.Join(visitedTables, " -> ")
 }
